+++ app/messages/table.ts
@@ -1,27 +1,27 @@
 import { DynamoDB } from 'aws-sdk';
-import { Hello } from './hello';
+import { Message } from './message';
 
-const TableName = process.env.HELLO_TABLE;
+const TableName = process.env.MESSAGES_TABLE;
 const client = new DynamoDB.DocumentClient({ region: process.env.REGION, apiVersion: '2012-08-10' });
 
-export class HelloTable {
+export class MessagesTable {
   constructor(private table: DynamoDB.DocumentClient) {}
 
-  add(Item: Hello): Promise<any> {
-    console.log('Adding hello', Item);
+  add(Item: Message): Promise<any> {
+    console.log('Adding message', Item);
     return this.table.put({ TableName, Item }).promise();
   }
 
-  async getAll(): Promise<Hello[]> {
-    console.log('Getting all hellos');
+  async getAll(): Promise<Message[]> {
+    console.log('Getting all messages');
     const res = await this.table.scan({ TableName }).promise();
-    return <Hello[]>res.Items;
+    return <Message[]>res.Items;
   }
 
   del(id: string): Promise<any> {
-    console.log('Deleting hello with id', id);
+    console.log('Deleting message with id', id);
     return this.table.delete({ TableName, Key: { id } }).promise();
   }
 }
 
-export const helloTable = new HelloTable(client);
+export const messagesTable = new MessagesTable(client);