+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## 2.0.0&nbsp;&nbsp;(2019-10-18)
+
+### Added
+
+- Added more complete swagger definitions for objects
+
+### Changed
+
+- Moved from confusing 'hello' model to 'messages'
+- Updated libraries
+
 ## 1.0.0&nbsp;&nbsp;(2019-08-24)
 
 ### Added